requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-pandas==0.6.7
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
idna==3.10
jinxed==1.3.0
numpy==2.1.2
pandas==2.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
redis==3.5.3
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
